First Files
Main is for the keypad. You can use it by sending serial signals, all you need to send is a 1-9 for the numbers, v for verb, n for noun, p for pro, k for key release etc. JSON_export is for exporting the values in the AGC JSON file, you need to edit the path in the script to yours to use it.